// src/app/modules/manufacturing/manufacturing-dashboard/manufacturing-summary-overview/models/manufacturing-summary.ts
export interface ManufacturedProductCountDTO {
    productType: string;
    count: number;
}

export interface ManufacturedProductStatusCountDTO {
    status: string;
    count: number;
}

export interface MonthlyManufacturedProductCountDTO {
    month: number;
    productType: string;
    count: number;
}